Combine store updaters for ease of review
